Move contract to separate module
